Refactor: "컴포넌트 분리"
한 페이지 다량의 코드를 분리
